Move Utils class inside utils package
