+++ app/src/test/java/io/horizontalsystems/bankwallet/core/EthInputParserTest.kt
@@ -1,5 +1,6 @@
 package io.horizontalsystems.bankwallet.core
 
+import io.horizontalsystems.bankwallet.core.utils.EthInputParser
 import org.junit.Assert.assertEquals
 import org.junit.Test
